Extract string commands and tests to separate files
